Changes to get ready for Python Package  (#4)
* add folders for ENUMs and COMs + implement COM.Adresse

* add initial pylint.yml

* add pylint to requirements.txt

* add enum Marktrolle

* update README.md (copied from BO4E-dotnet repo)

* reformat all files using black

* explicitly lint certain directories

* Revert "explicitly lint certain directories"

This reverts commit 37cfefd5b3511104bf9666326565036a9c869ab6.

* Set version >= 3.8 in readme (instead of >)

* Add tox.ini

In this file we describe each virtual environment and all commands which should be executed in each venv. It still needs some polishing.

* Split requirements into different files

It helps to more clearly separate the dependencies for the different virtual environments for testing, documentation and actual development.

* Change project structure to src layout

* Add __init__.py files into each folder to get a working import structure

* Delete __init__.py caused by src layout

* Add versioneer to repository

versioneer takes care of a simple and neat versioning of our package

* Split most of the information into setup.cfg

Use the configuration file setup.cfg for all parameters used in setup.py.

* Convert README from markdown to reStructuredText format

This conversion is motivated by Sphinx. Later we can also change back to markdown. But .rst files are default in Sphinx.

* Add AUTHORS.rst and CHANGELOG.rst

* Rename LICENSE to LICENSE.rst

* Add docs folder with the initial documentation content + logo

* Add sphinx_rtd_theme requirement

realises a nicer looking documentation layout and is the default for documetations on readthedocs

* Add *.todo to .gitignore

* Comment out coverage-report evironment

this is not ready yet. But will come handy in the future, so I keep it there as a reminder

* Add .gitattributes which is coming from `versioneer install`

* Switch test action to tox commands

* Rename unittests to tests and fix syntax error

* Remove unused json import

* Replace too long line by multiline string

* Remove unused import `from enum import Enum`

* Replace pylint with flake8

All big python projects are using flake8 instead of pylint.

* Add doctests and flake8 config into tox.ini

* put import commands to the top

* Remove three `-` signs cause the line was too long

* rename file to linting.yml and replace pylint with flake8

Co-authored-by: Konstantin KLein <konstantin.klein@hochfrequenz.de>