Change construction for optional attributes
If you set an attribute to `init=False` you will get an issue in the deserialisation process. The json which is created from an object which does not contain an optional attribute won't contain it. But during the deserialisation process **all** potential attributes are required to be there. So we set all optional attributes per default to None.