Update address requirements
The minimal address in Germany is only a zip code and city  (for example for institutions like universities or big companies or postfächer without a number but a special zip code) 
A street always needs a hausnummer and vice versa.
Postfächer don't have a street, but a number OR a zipcode.