🩹 Add BoTyp.GESCHAEFTSOBJEKT and set it to default
The BoTyp can be used as a descriminator in databases.
But the Geschaeftsobjekt itself has no BoTyp in the documentation.
To realize a good table inheritance you have to define a table for each class and **each** class have to identified by a discriminator. Therefore we added the BoTyp.GESCHAEFTSOBJEKT to the BoTyp enumeration.